pyproject.toml
[project]
name = "labelanything"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "colorlog>=6.9.0",
    "dropblock>=0.3.0",
    "easydict>=1.13",
    "einops>=0.8.1",
    "matplotlib>=3.10.1",
    "opencv-python>=4.11.0.86",
    "optuna>=4.3.0",
    "pandas>=2.2.3",
    "pycocotools>=2.0.8",
    "pyyaml>=6.0.2",
    "ruamel-yaml>=0.18.10",
    "scipy>=1.15.2",
    "timm>=1.0.15",
    "torch>=2.7.0",
    "torchmetrics>=1.7.1",
    "torchvision>=0.22.0",
    "transformers>=4.51.3",
    "wandb>=0.19.10",
]
